start xow.service on dongle plugin
I encountered the `Mt76Exception: Failed to load firmware` error quite a few times. xow.service is stopped and needs a restart as stated in the Troubleshooting section of the README.md. To make my life easier i added a udev rule to do just that for me when i plugged the dongle back in after uplugging it.
Should start/restart the xow.service if the service is not running or is in a `failed` state due to previous ABRT signal.
I think it would be a simple change to the udev.rules to make that the default behaviour but maybe a manual restart is desired?

Anyways, thanks a lot for your work! I'm quite happy with xow!