Load the firmware file from disk
Load the firmware file from disk, instead of the current approach where
we embed the binary within the executable.

In case the file is missing, print instructions how to fetch it.
Explicitly mention the Microsoft Terms of Use.

With this at hand, it is the user's responsibility to ensure licensing
and terms of use compliance.

v2:
 - Create DATADIR during install
 - Don't skip whitespace from the input stream

v3:
 - Use /lib/firmware/xow_dongle.bin
 - Keep the firmware blob local
 - Use uint8_t array

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>